IoT2_Start/Azure VM/requirements.txt
blinker==1.7.0
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
Flask==3.0.2
fonttools==4.49.0
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
numpy==1.26.4
packaging==23.2
paho-mqtt==2.0.0
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.9.0.post0
six==1.16.0
Werkzeug==3.0.1
